battery: refactor sysfs handler into util module
