+++ src/util/mod.rs
@@ -1,1 +1,2 @@
 pub mod error;
+pub mod sysfs;
